// temp/cms/models/com_ever_edu_cms_asgmt_dto_req_AsgmtSubmitSaveReqDto$onUser.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type com_ever_edu_cms_asgmt_dto_req_AsgmtSubmitSaveReqDto$onUser = {
    /**
     * 과제 ID
     */
    asgmtId?: number;
    /**
     * 과제 제출 제목
     */
    submitTitle?: string;
    /**
     * 과제 답변 텍스트
     */
    answerText?: string | null;
    /**
     * 과제 답변 파일ID
     */
    answerFileId?: number | null;
};