// temp/cms/models/com_ever_edu_cms_video_dto_req_WatchLogSearchDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type com_ever_edu_cms_video_dto_req_WatchLogSearchDto = {
    /**
     * 차수 ID
     */
    courseSequenceId?: number;
    /**
     * 코스 ID
     */
    courseId?: number;
    /**
     * 커리큘럼 ID
     */
    curriculumId?: number;
    /**
     * 모듈 ID
     */
    moduleId?: number;
    /**
     * 레슨 ID
     */
    lessonId?: number;
    /**
     * 콘텐츠 UUID
     */
    contentUuid?: string;
};