// temp/cms/models/org_springframework_data_domain_PageCom_ever_edu_cms_curriculum_dto_res_CurriculumSearchResDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { com_ever_edu_cms_curriculum_dto_res_CurriculumSearchResDto } from './com_ever_edu_cms_curriculum_dto_res_CurriculumSearchResDto';
import type { org_springdoc_core_converters_models_SortObject } from './org_springdoc_core_converters_models_SortObject';
import type { PageableObject } from './PageableObject';
export type org_springframework_data_domain_PageCom_ever_edu_cms_curriculum_dto_res_CurriculumSearchResDto = {
    totalPages?: number;
    totalElements?: number;
    size?: number;
    content?: Array<com_ever_edu_cms_curriculum_dto_res_CurriculumSearchResDto>;
    number?: number;
    sort?: Array<org_springdoc_core_converters_models_SortObject>;
    numberOfElements?: number;
    pageable?: PageableObject;
    first?: boolean;
    last?: boolean;
    empty?: boolean;
};